Hawk: split Launcher into abstract + batch halves
