+++ solutions/Hawk/org.hawk.ttc2018/Hawk TTC2018 (size 1).launch
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
 <launchConfiguration type="org.eclipse.jdt.launching.localJavaApplication">
 <listAttribute key="org.eclipse.debug.core.MAPPED_RESOURCE_PATHS">
-<listEntry value="/org.hawk.ttc2018/src/org/hawk/ttc2018/Launcher.java"/>
+<listEntry value="/org.hawk.ttc2018/src/org/hawk/ttc2018/BatchLauncher.java"/>
 </listAttribute>
 <listAttribute key="org.eclipse.debug.core.MAPPED_RESOURCE_TYPES">
 <listEntry value="1"/>
@@ -14,7 +14,7 @@
 <mapEntry key="Sequences" value="20"/>
 </mapAttribute>
 <booleanAttribute key="org.eclipse.jdt.launching.ATTR_USE_START_ON_FIRST_THREAD" value="true"/>
-<stringAttribute key="org.eclipse.jdt.launching.MAIN_TYPE" value="org.hawk.ttc2018.Launcher"/>
+<stringAttribute key="org.eclipse.jdt.launching.MAIN_TYPE" value="org.hawk.ttc2018.BatchLauncher"/>
 <stringAttribute key="org.eclipse.jdt.launching.PROJECT_ATTR" value="org.hawk.ttc2018"/>
 <stringAttribute key="org.eclipse.jdt.launching.VM_ARGUMENTS" value="-Xmx8g -Xms8g"/>
 </launchConfiguration>
